+++ src/CarFuelMonitor.java
@@ -1,130 +1,59 @@
-import components.simplewriter.SimpleWriter;
-import components.simplewriter.SimpleWriter1L;
 
 /**
+ * CarFuelMonitorKernel provides an way to monitor and manage the fuel level of
+ * a car. It allows adding fuel, checking whether the car is out of fuel, and
+ * retrieving the current fuel level.
  *
- * {@code CarFuelMontor} represented as a {@code double} with implementations of
- * primary methods.
+ * <p>
+ * The fuel level is represented as an double, where the value 0 indicates that
+ * the car is out of fuel. Any positive value indicates the amount of fuel
+ * currently available. The interface assumes a non-negative fuel level at all
+ * times, i.e., the fuel level cannot drop below zero.
+ * </p>
  *
- * @convention
- *
- *             <pre>
- * [all characters of $this.rep are '0' through '9']  and
- * [$this.rep does not start with '0']
- *             </pre>
- *
- * @correspondence
- *
- *                 <pre>
- * this = [if $this.rep = "" then 0
- *         else the decimal number whose ordinary depiction is $this.rep]
- *                 </pre>
- *
- * @author Jay Patel
- *
- **/
-public class CarFuelMonitor {
+ * @initially <pre>
+ *     this.getCurrentFuelLevel() = 0
+ * </pre>
+ */
+public interface CarFuelMonitor extends CarFuelMonitorKernel {
 
     /**
-     * Fuel level of the car, representation of {@code this}.
-     */
-    private double fuelLevel;
-
-    /**
-     * No-argument constructor. Initializes fuel level to 0.
-     */
-    public CarFuelMonitor() {
-        this.fuelLevel = 0.0;
-    }
-
-    /**
-     * Constructor with initial fuel level.
+     * Simulates driving by decreasing the fuel level based on fuel efficiency.
+     * Calculates and returns the distance covered.
      *
-     * @param initialFuelLevel
-     *            initial fuel level
+     * @param fuelUsed
+     *            the amount of fuel intended to be used
+     * @param efficiency
+     *            the fuel efficiency (fuel consumed per unit distance)
+     * @return the actual distance covered based on fuel used
+     * @requires fuelUsed >= 0 and efficiency > 0
+     * @ensures this.fuelLevel = #this.fuelLevel - [actual fuel used]
      */
-    public CarFuelMonitor(double initialFuelLevel) {
-        assert initialFuelLevel >= 0 : "Violation of: initialFuelLevel >= 0";
-        this.fuelLevel = initialFuelLevel;
-    }
-
-    /*
-     * Kernel methods
-     * -----------------------------------------------------------
-     */
+    double drive(double fuelUsed, double efficiency);
 
     /**
-     * Adds fuel to the car.
+     * Refuels the car with the specified amount of fuel, but does not exceed
+     * max capacity.
      *
      * @param amount
      *            the amount of fuel to add
+     * @param maxCapacity
+     *            the max fuel capacity of the car
+     * @return the actual amount of fuel added
+     * @requires amount >= 0 and maxCapacity >= this.fuelLevel
+     * @ensures this.fuelLevel = min(#this.fuelLevel + amount, maxCapacity)
      */
-    public void addFuel(double amount) {
-        assert amount >= 0 : "Violation of: amount >= 0";
-
-        this.fuelLevel += amount;
-    }
+    double refuel(double amount, double maxCapacity);
 
     /**
-     * Checks if the car is out of fuel.
+     * Checks if the car has enough fuel to reach the destination.
      *
-     * @return true if fuel level is 0, false otherwise
+     * @param distanceToDestination
+     *            the distance to the destination
+     * @param fuelEfficiency
+     *            the car's fuel efficiency (e.g., gallons per mile)
+     * @return true if there is not enough fuel, false otherwise
      */
-    public boolean isOutOfFuel() {
-        return this.fuelLevel == 0;
-    }
-
-    /**
-     * Gets the current fuel level.
-     *
-     * @return the current fuel level
-     */
-    public double getFuelLevel() {
-        return this.fuelLevel;
-    }
-
-    /**
-     * Main method.
-     *
-     * @param args
-     *            command-line arguments
-     */
-    public static void main(String[] args) {
-        CarFuelMonitor car = new CarFuelMonitor();
-
-        SimpleWriter out = new SimpleWriter1L();
-        car.addFuel(10.5);
-        out.println("Fuel added. Current fuel level: " + car.getFuelLevel()
-                + " liters");
-
-        if (car.isOutOfFuel()) {
-            out.println("The car is out of fuel.");
-        } else {
-            out.println("The car has fuel.");
-        }
-
-        car.addFuel(5.0);
-        out.println("More fuel added. Current fuel level: " + car.getFuelLevel()
-                + " liters");
-
-        car.addFuel(car.getFuelLevel());
-        out.println("All fuel consumed. Current fuel level: "
-                + car.getFuelLevel() + " liters");
-
-        if (car.isOutOfFuel()) {
-            out.println("The car is out of fuel.");
-        }
-        car.addFuel(10.0);
-        System.out.println(
-                "Current fuel level: " + car.getFuelLevel() + " liters");
-
-        if (car.isOutOfFuel()) {
-            System.out.println("The car is out of fuel.");
-        } else {
-            System.out.println("The car has fuel.");
-        }
-
-        out.close();
-    }
+    boolean isLowFuel(double distanceToDestination, double fuelEfficiency);
 
 }
